+++ tests/testthat/test-ComputeLayout.R
@@ -100,7 +100,7 @@ for (assay_version in c("v3", "v5")) {
     expect_s3_class(cg_layout@layout[["custom"]], "data.frame")
     expect_equal(colnames(cg_layout@layout[["custom"]]), c("x", "y", "z"))
     expect_lt(.row_names_info(cg_layout@layout[["custom"]]), 0L)
-    expect_equal(SeuratObject::Cells(cg_layout), node_names)
+    expect_equal(CellGraphData(cg_layout, slot = "nodes"), node_names)
   })
 
   test_that("custom layout coordinates are matched to nodes by name", {
@@ -115,7 +115,7 @@ for (assay_version in c("v3", "v5")) {
       custom_layout_function = shuffled_fkn,
       custom_layout_function_args = list(dim = 3)
     )
-    expect_equal(SeuratObject::Cells(cg_layout), node_names)
+    expect_equal(CellGraphData(cg_layout, slot = "nodes"), node_names)
     expect_lt(.row_names_info(cg_layout@layout[["custom"]]), 0L)
     expect_equal(
       as.numeric(as.matrix(cg_layout@layout[["custom"]])),
@@ -128,7 +128,7 @@ for (assay_version in c("v3", "v5")) {
       custom_layout_function = unnamed_fkn,
       custom_layout_function_args = list(dim = 3)
     )
-    expect_equal(SeuratObject::Cells(cg_layout), node_names)
+    expect_equal(CellGraphData(cg_layout, slot = "nodes"), node_names)
     expect_lt(.row_names_info(cg_layout@layout[["custom"]]), 0L)
     expect_equal(
       as.numeric(as.matrix(cg_layout@layout[["custom"]])),